pyproject.toml
[project]
name = "yeogida-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Sihyun Kim",email = "k4sud0n@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "django-ninja-jwt (>=5.3.7,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
pre-commit = "^4.2.0"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
